refactor: delete unused + nodes-and-edges => database
