--- src/index.test.ts
@@ -1,7 +1,0 @@
-import { describe, expect, it } from "vitest"
-
-describe("sum test", () => {
-  it("adds 1 + 2 to equal 3", () => {
-    expect(1 + 2).toBe(3)
-  })
-})
