--- src/lib/message.ts
@@ -1,1 +1,0 @@
-export const message = "hello from $lib/message"
