--- test-results/.last-run.json
@@ -1,4 +1,0 @@
-{
-  "status": "failed",
-  "failedTests": ["5f7b6aabace88f74752a-1b97bc9f8fb61ed19afc"]
-}
